@@ -1,7 +1,7 @@
 source 'https://rubygems.org'
-ruby '2.3.1'
+ruby '2.5.0'
 
-gem 'rails', '4.2.6'
+gem 'rails', '5.2.1'
 
 # Use postgresql as the database for Active Record
 gem 'pg', '~> 0.15'
