@@ -1,4 +1,4 @@
-class CreateCollections < ActiveRecord::Migration
+class CreateCollections < ActiveRecord::Migration[5.2]
   def change
     create_table :collections do |t|
       t.string :name,       null: false
