@@ -1,4 +1,4 @@
-class RemoveBodyFromIssues < ActiveRecord::Migration
+class RemoveBodyFromIssues < ActiveRecord::Migration[5.2]
   def change
     remove_column :issues, :body
   end
