@@ -1,4 +1,4 @@
-class AddUserToCollection < ActiveRecord::Migration
+class AddUserToCollection < ActiveRecord::Migration[5.2]
   def change
     add_reference :collections, :user, index: true, foreign_key: true
   end
