@@ -1,4 +1,4 @@
-class CreateLabels < ActiveRecord::Migration
+class CreateLabels < ActiveRecord::Migration[5.2]
   def change
     create_table :labels do |t|
       t.string :name
