@@ -1,4 +1,4 @@
-class RemoveLabelsFromIssue < ActiveRecord::Migration
+class RemoveLabelsFromIssue < ActiveRecord::Migration[5.2]
   def change
     remove_column :issues, :labels
   end
