@@ -1,4 +1,4 @@
-class AddSlugToProject < ActiveRecord::Migration
+class AddSlugToProject < ActiveRecord::Migration[5.2]
   def change
     add_column :projects, :slug, :string
     add_index :projects, :slug, unique: true
