@@ -1,4 +1,4 @@
-class ChangeNameToUrlOnOrganization < ActiveRecord::Migration
+class ChangeNameToUrlOnOrganization < ActiveRecord::Migration[5.2]
   def change
     rename_column :organizations, :name, :url
   end
