@@ -1,4 +1,4 @@
-class DropDelayed < ActiveRecord::Migration
+class DropDelayed < ActiveRecord::Migration[5.2]
   def change
     drop_table :delayed_jobs
   end
