@@ -1,4 +1,4 @@
-class AddIndexToProjectUrl < ActiveRecord::Migration
+class AddIndexToProjectUrl < ActiveRecord::Migration[5.2]
   def change
     add_index :projects, :url
   end
