@@ -12,12 +12,12 @@ class ApplicationController < ActionController::Base
     orgs_for_search = orgs.map { |org| "org:" + org.name }.uniq
     repos_for_search = repos.map { |repo| "repo:" + repo.full_name }.uniq
     help_wanted_query = "type:issue state:open label:\"help wanted\"" + orgs_for_search.join(" ") + " " + repos_for_search.join(" ")
-    help_wanted_result = @github_api.search_issues(help_wanted_query, { sort: "updated", per_page: 10 })
+    help_wanted_result = @github_api.search_issues(help_wanted_query, { sort: "updated", per_page: 100 })
     @issues = help_wanted_result.items.map { |github_issue| Issue.new github_issue }
 
     if featured.present?
       featured_query = "type:issue state:open label:\"help wanted\" label:\"#{featured}\"" + orgs_for_search.join(" ") + " " + repos_for_search.join(" ")
-      featured_result = @github_api.search_issues(featured_query, { sort: "updated", per_page: 10 })
+      featured_result = @github_api.search_issues(featured_query, { sort: "updated", per_page: 100 })
       featured_issues = featured_result.items.map { |github_issue| Issue.new github_issue }
       @issues = featured_issues + @issues
       @issues = @issues.uniq { |issue| issue.url }
